removed individual bookviews and replaced them with one reflowable. still a lot more work to be done, but the first step was to migrate both into one
